Rename store integration package Just change the base package of Store integration.
Previously it was:

- org.spin.model
- org.spin.util
- org.spin.util.support

Now is:

- org.spin.store.model
- org.spin.store.util
- org.spin.store.util.support
