--- org.spin.store/src/main/java/org/spin/model/I_C_PaymentMethodAllocation.java
@@ -1,188 +1,0 @@
-/******************************************************************************
- * Product: ADempiere ERP & CRM Smart Business Solution                       *
- * Copyright (C) 2006-2017 ADempiere Foundation, All Rights Reserved.         *
- * This program is free software, you can redistribute it and/or modify it    *
- * under the terms version 2 of the GNU General Public License as published   *
- * or (at your option) any later version.										*
- * by the Free Software Foundation. This program is distributed in the hope   *
- * that it will be useful, but WITHOUT ANY WARRANTY, without even the implied *
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.           *
- * See the GNU General Public License for more details.                       *
- * You should have received a copy of the GNU General Public License along    *
- * with this program, if not, write to the Free Software Foundation, Inc.,    *
- * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA.                     *
- * For the text or an alternative of this public license, you may reach us    *
- * or via info@adempiere.net or http://www.adempiere.net/license.html         *
- *****************************************************************************/
-package org.spin.model;
-
-import java.math.BigDecimal;
-import java.sql.Timestamp;
-import org.compiere.model.*;
-import org.compiere.util.KeyNamePair;
-
-/** Generated Interface for C_PaymentMethodAllocation
- *  @author Adempiere (generated) 
- *  @version Release 3.9.3
- */
-public interface I_C_PaymentMethodAllocation 
-{
-
-    /** TableName=C_PaymentMethodAllocation */
-    public static final String Table_Name = "C_PaymentMethodAllocation";
-
-    /** AD_Table_ID=54775 */
-    public static final int Table_ID = MTable.getTable_ID(Table_Name);
-
-    KeyNamePair Model = new KeyNamePair(Table_ID, Table_Name);
-
-    /** AccessLevel = 3 - Client - Org 
-     */
-    BigDecimal accessLevel = BigDecimal.valueOf(3);
-
-    /** Load Meta Data */
-
-    /** Column name AD_Client_ID */
-    public static final String COLUMNNAME_AD_Client_ID = "AD_Client_ID";
-
-	/** Get Client.
-	  * Client/Tenant for this installation.
-	  */
-	public int getAD_Client_ID();
-
-    /** Column name AD_Org_ID */
-    public static final String COLUMNNAME_AD_Org_ID = "AD_Org_ID";
-
-	/** Set Organization.
-	  * Organizational entity within client
-	  */
-	public void setAD_Org_ID (int AD_Org_ID);
-
-	/** Get Organization.
-	  * Organizational entity within client
-	  */
-	public int getAD_Org_ID();
-
-    /** Column name C_PaymentMethodAllocation_ID */
-    public static final String COLUMNNAME_C_PaymentMethodAllocation_ID = "C_PaymentMethodAllocation_ID";
-
-	/** Set Store Payment Method	  */
-	public void setC_PaymentMethodAllocation_ID (int C_PaymentMethodAllocation_ID);
-
-	/** Get Store Payment Method	  */
-	public int getC_PaymentMethodAllocation_ID();
-
-    /** Column name C_PaymentMethod_ID */
-    public static final String COLUMNNAME_C_PaymentMethod_ID = "C_PaymentMethod_ID";
-
-	/** Set Store Payment Method.
-	  * Payment Methods allowed for Store
-	  */
-	public void setC_PaymentMethod_ID (int C_PaymentMethod_ID);
-
-	/** Get Store Payment Method.
-	  * Payment Methods allowed for Store
-	  */
-	public int getC_PaymentMethod_ID();
-
-	public org.spin.model.I_C_PaymentMethod getC_PaymentMethod() throws RuntimeException;
-
-    /** Column name Created */
-    public static final String COLUMNNAME_Created = "Created";
-
-	/** Get Created.
-	  * Date this record was created
-	  */
-	public Timestamp getCreated();
-
-    /** Column name CreatedBy */
-    public static final String COLUMNNAME_CreatedBy = "CreatedBy";
-
-	/** Get Created By.
-	  * User who created this records
-	  */
-	public int getCreatedBy();
-
-    /** Column name Description */
-    public static final String COLUMNNAME_Description = "Description";
-
-	/** Set Description.
-	  * Optional short description of the record
-	  */
-	public void setDescription (String Description);
-
-	/** Get Description.
-	  * Optional short description of the record
-	  */
-	public String getDescription();
-
-    /** Column name IsActive */
-    public static final String COLUMNNAME_IsActive = "IsActive";
-
-	/** Set Active.
-	  * The record is active in the system
-	  */
-	public void setIsActive (boolean IsActive);
-
-	/** Get Active.
-	  * The record is active in the system
-	  */
-	public boolean isActive();
-
-    /** Column name Note */
-    public static final String COLUMNNAME_Note = "Note";
-
-	/** Set Note.
-	  * Optional additional user defined information
-	  */
-	public void setNote (String Note);
-
-	/** Get Note.
-	  * Optional additional user defined information
-	  */
-	public String getNote();
-
-    /** Column name Updated */
-    public static final String COLUMNNAME_Updated = "Updated";
-
-	/** Get Updated.
-	  * Date this record was updated
-	  */
-	public Timestamp getUpdated();
-
-    /** Column name UpdatedBy */
-    public static final String COLUMNNAME_UpdatedBy = "UpdatedBy";
-
-	/** Get Updated By.
-	  * User who updated this records
-	  */
-	public int getUpdatedBy();
-
-    /** Column name UUID */
-    public static final String COLUMNNAME_UUID = "UUID";
-
-	/** Set Immutable Universally Unique Identifier.
-	  * Immutable Universally Unique Identifier
-	  */
-	public void setUUID (String UUID);
-
-	/** Get Immutable Universally Unique Identifier.
-	  * Immutable Universally Unique Identifier
-	  */
-	public String getUUID();
-
-    /** Column name W_Store_ID */
-    public static final String COLUMNNAME_W_Store_ID = "W_Store_ID";
-
-	/** Set Web Store.
-	  * A Web Store of the Client
-	  */
-	public void setW_Store_ID (int W_Store_ID);
-
-	/** Get Web Store.
-	  * A Web Store of the Client
-	  */
-	public int getW_Store_ID();
-
-	public org.compiere.model.I_W_Store getW_Store() throws RuntimeException;
-}
