--- org.spin.store/src/main/java/org/spin/model/X_W_Category.java
@@ -1,245 +1,0 @@
-/******************************************************************************
- * Product: ADempiere ERP & CRM Smart Business Solution                       *
- * Copyright (C) 2006-2017 ADempiere Foundation, All Rights Reserved.         *
- * This program is free software, you can redistribute it and/or modify it    *
- * under the terms version 2 of the GNU General Public License as published   *
- * or (at your option) any later version.										*
- * by the Free Software Foundation. This program is distributed in the hope   *
- * that it will be useful, but WITHOUT ANY WARRANTY, without even the implied *
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.           *
- * See the GNU General Public License for more details.                       *
- * You should have received a copy of the GNU General Public License along    *
- * with this program, if not, write to the Free Software Foundation, Inc.,    *
- * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA.                     *
- * For the text or an alternative of this public license, you may reach us    *
- * or via info@adempiere.net or http://www.adempiere.net/license.html         *
- *****************************************************************************/
-/** Generated Model - DO NOT CHANGE */
-package org.spin.model;
-
-import java.sql.ResultSet;
-import java.util.Properties;
-import org.compiere.model.*;
-import org.compiere.util.KeyNamePair;
-
-/** Generated Model for W_Category
- *  @author Adempiere (generated) 
- *  @version Release 3.9.3 - $Id$ */
-public class X_W_Category extends PO implements I_W_Category, I_Persistent 
-{
-
-	/**
-	 *
-	 */
-	private static final long serialVersionUID = 20201102L;
-
-    /** Standard Constructor */
-    public X_W_Category (Properties ctx, int W_Category_ID, String trxName)
-    {
-      super (ctx, W_Category_ID, trxName);
-      /** if (W_Category_ID == 0)
-        {
-			setIsDefault (false);
-			setName (null);
-			setValue (null);
-			setW_Category_ID (0);
-        } */
-    }
-
-    /** Load Constructor */
-    public X_W_Category (Properties ctx, ResultSet rs, String trxName)
-    {
-      super (ctx, rs, trxName);
-    }
-
-    /** AccessLevel
-      * @return 3 - Client - Org 
-      */
-    protected int get_AccessLevel()
-    {
-      return accessLevel.intValue();
-    }
-
-    /** Load Meta Data */
-    protected POInfo initPO (Properties ctx)
-    {
-      POInfo poi = POInfo.getPOInfo (ctx, Table_ID, get_TrxName());
-      return poi;
-    }
-
-    public String toString()
-    {
-      StringBuffer sb = new StringBuffer ("X_W_Category[")
-        .append(get_ID()).append("]");
-      return sb.toString();
-    }
-
-	/** Set Description.
-		@param Description 
-		Optional short description of the record
-	  */
-	public void setDescription (String Description)
-	{
-		set_Value (COLUMNNAME_Description, Description);
-	}
-
-	/** Get Description.
-		@return Optional short description of the record
-	  */
-	public String getDescription () 
-	{
-		return (String)get_Value(COLUMNNAME_Description);
-	}
-
-	/** Set Default.
-		@param IsDefault 
-		Default value
-	  */
-	public void setIsDefault (boolean IsDefault)
-	{
-		set_Value (COLUMNNAME_IsDefault, Boolean.valueOf(IsDefault));
-	}
-
-	/** Get Default.
-		@return Default value
-	  */
-	public boolean isDefault () 
-	{
-		Object oo = get_Value(COLUMNNAME_IsDefault);
-		if (oo != null) 
-		{
-			 if (oo instanceof Boolean) 
-				 return ((Boolean)oo).booleanValue(); 
-			return "Y".equals(oo);
-		}
-		return false;
-	}
-
-	/** Set Featured in Web Store.
-		@param IsWebStoreFeatured 
-		If selected, the product is displayed in the initial or any empty search
-	  */
-	public void setIsWebStoreFeatured (boolean IsWebStoreFeatured)
-	{
-		set_Value (COLUMNNAME_IsWebStoreFeatured, Boolean.valueOf(IsWebStoreFeatured));
-	}
-
-	/** Get Featured in Web Store.
-		@return If selected, the product is displayed in the initial or any empty search
-	  */
-	public boolean isWebStoreFeatured () 
-	{
-		Object oo = get_Value(COLUMNNAME_IsWebStoreFeatured);
-		if (oo != null) 
-		{
-			 if (oo instanceof Boolean) 
-				 return ((Boolean)oo).booleanValue(); 
-			return "Y".equals(oo);
-		}
-		return false;
-	}
-
-	/** Set Name.
-		@param Name 
-		Alphanumeric identifier of the entity
-	  */
-	public void setName (String Name)
-	{
-		set_Value (COLUMNNAME_Name, Name);
-	}
-
-	/** Get Name.
-		@return Alphanumeric identifier of the entity
-	  */
-	public String getName () 
-	{
-		return (String)get_Value(COLUMNNAME_Name);
-	}
-
-    /** Get Record ID/ColumnName
-        @return ID/ColumnName pair
-      */
-    public KeyNamePair getKeyNamePair() 
-    {
-        return new KeyNamePair(get_ID(), getName());
-    }
-
-	/** Set Immutable Universally Unique Identifier.
-		@param UUID 
-		Immutable Universally Unique Identifier
-	  */
-	public void setUUID (String UUID)
-	{
-		set_Value (COLUMNNAME_UUID, UUID);
-	}
-
-	/** Get Immutable Universally Unique Identifier.
-		@return Immutable Universally Unique Identifier
-	  */
-	public String getUUID () 
-	{
-		return (String)get_Value(COLUMNNAME_UUID);
-	}
-
-	/** Set Search Key.
-		@param Value 
-		Search key for the record in the format required - must be unique
-	  */
-	public void setValue (String Value)
-	{
-		set_Value (COLUMNNAME_Value, Value);
-	}
-
-	/** Get Search Key.
-		@return Search key for the record in the format required - must be unique
-	  */
-	public String getValue () 
-	{
-		return (String)get_Value(COLUMNNAME_Value);
-	}
-
-	/** Set Product Store Group.
-		@param W_Category_ID Product Store Group	  */
-	public void setW_Category_ID (int W_Category_ID)
-	{
-		if (W_Category_ID < 1) 
-			set_ValueNoCheck (COLUMNNAME_W_Category_ID, null);
-		else 
-			set_ValueNoCheck (COLUMNNAME_W_Category_ID, Integer.valueOf(W_Category_ID));
-	}
-
-	/** Get Product Store Group.
-		@return Product Store Group	  */
-	public int getW_Category_ID () 
-	{
-		Integer ii = (Integer)get_Value(COLUMNNAME_W_Category_ID);
-		if (ii == null)
-			 return 0;
-		return ii.intValue();
-	}
-
-	public org.spin.model.I_W_Category getW_Category_Parent() throws RuntimeException
-    {
-		return (org.spin.model.I_W_Category)MTable.get(getCtx(), org.spin.model.I_W_Category.Table_Name)
-			.getPO(getW_Category_Parent_ID(), get_TrxName());	}
-
-	/** Set Product Store Group Parent.
-		@param W_Category_Parent_ID Product Store Group Parent	  */
-	public void setW_Category_Parent_ID (int W_Category_Parent_ID)
-	{
-		if (W_Category_Parent_ID < 1) 
-			set_Value (COLUMNNAME_W_Category_Parent_ID, null);
-		else 
-			set_Value (COLUMNNAME_W_Category_Parent_ID, Integer.valueOf(W_Category_Parent_ID));
-	}
-
-	/** Get Product Store Group Parent.
-		@return Product Store Group Parent	  */
-	public int getW_Category_Parent_ID () 
-	{
-		Integer ii = (Integer)get_Value(COLUMNNAME_W_Category_Parent_ID);
-		if (ii == null)
-			 return 0;
-		return ii.intValue();
-	}
-}
