--- org.spin.store/src/main/java/org/spin/model/X_W_DeliveryViaRuleAllocation.java
@@ -1,287 +1,0 @@
-/******************************************************************************
- * Product: ADempiere ERP & CRM Smart Business Solution                       *
- * Copyright (C) 2006-2017 ADempiere Foundation, All Rights Reserved.         *
- * This program is free software, you can redistribute it and/or modify it    *
- * under the terms version 2 of the GNU General Public License as published   *
- * or (at your option) any later version.										*
- * by the Free Software Foundation. This program is distributed in the hope   *
- * that it will be useful, but WITHOUT ANY WARRANTY, without even the implied *
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.           *
- * See the GNU General Public License for more details.                       *
- * You should have received a copy of the GNU General Public License along    *
- * with this program, if not, write to the Free Software Foundation, Inc.,    *
- * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA.                     *
- * For the text or an alternative of this public license, you may reach us    *
- * or via info@adempiere.net or http://www.adempiere.net/license.html         *
- *****************************************************************************/
-/** Generated Model - DO NOT CHANGE */
-package org.spin.model;
-
-import java.sql.ResultSet;
-import java.util.Properties;
-import org.compiere.model.*;
-
-/** Generated Model for W_DeliveryViaRuleAllocation
- *  @author Adempiere (generated) 
- *  @version Release 3.9.3 - $Id$ */
-public class X_W_DeliveryViaRuleAllocation extends PO implements I_W_DeliveryViaRuleAllocation, I_Persistent 
-{
-
-	/**
-	 *
-	 */
-	private static final long serialVersionUID = 20201019L;
-
-    /** Standard Constructor */
-    public X_W_DeliveryViaRuleAllocation (Properties ctx, int W_DeliveryViaRuleAllocation_ID, String trxName)
-    {
-      super (ctx, W_DeliveryViaRuleAllocation_ID, trxName);
-      /** if (W_DeliveryViaRuleAllocation_ID == 0)
-        {
-			setDeliveryViaRule (null);
-// D
-			setW_DeliveryViaRuleAllocation_ID (0);
-			setW_Store_ID (0);
-        } */
-    }
-
-    /** Load Constructor */
-    public X_W_DeliveryViaRuleAllocation (Properties ctx, ResultSet rs, String trxName)
-    {
-      super (ctx, rs, trxName);
-    }
-
-    /** AccessLevel
-      * @return 3 - Client - Org 
-      */
-    protected int get_AccessLevel()
-    {
-      return accessLevel.intValue();
-    }
-
-    /** Load Meta Data */
-    protected POInfo initPO (Properties ctx)
-    {
-      POInfo poi = POInfo.getPOInfo (ctx, Table_ID, get_TrxName());
-      return poi;
-    }
-
-    public String toString()
-    {
-      StringBuffer sb = new StringBuffer ("X_W_DeliveryViaRuleAllocation[")
-        .append(get_ID()).append("]");
-      return sb.toString();
-    }
-
-	/** DeliveryViaRule AD_Reference_ID=152 */
-	public static final int DELIVERYVIARULE_AD_Reference_ID=152;
-	/** Pickup = P */
-	public static final String DELIVERYVIARULE_Pickup = "P";
-	/** Delivery = D */
-	public static final String DELIVERYVIARULE_Delivery = "D";
-	/** Shipper = S */
-	public static final String DELIVERYVIARULE_Shipper = "S";
-	/** Set Delivery Via.
-		@param DeliveryViaRule 
-		How the order will be delivered
-	  */
-	public void setDeliveryViaRule (String DeliveryViaRule)
-	{
-
-		set_Value (COLUMNNAME_DeliveryViaRule, DeliveryViaRule);
-	}
-
-	/** Get Delivery Via.
-		@return How the order will be delivered
-	  */
-	public String getDeliveryViaRule () 
-	{
-		return (String)get_Value(COLUMNNAME_DeliveryViaRule);
-	}
-
-	/** Set Description.
-		@param Description 
-		Optional short description of the record
-	  */
-	public void setDescription (String Description)
-	{
-		set_Value (COLUMNNAME_Description, Description);
-	}
-
-	/** Get Description.
-		@return Optional short description of the record
-	  */
-	public String getDescription () 
-	{
-		return (String)get_Value(COLUMNNAME_Description);
-	}
-
-	/** FreightCostRule AD_Reference_ID=153 */
-	public static final int FREIGHTCOSTRULE_AD_Reference_ID=153;
-	/** Freight included = I */
-	public static final String FREIGHTCOSTRULE_FreightIncluded = "I";
-	/** Fix price = F */
-	public static final String FREIGHTCOSTRULE_FixPrice = "F";
-	/** Calculated = C */
-	public static final String FREIGHTCOSTRULE_Calculated = "C";
-	/** Line = L */
-	public static final String FREIGHTCOSTRULE_Line = "L";
-	/** Set Freight Cost Rule.
-		@param FreightCostRule 
-		Method for charging Freight
-	  */
-	public void setFreightCostRule (String FreightCostRule)
-	{
-
-		set_Value (COLUMNNAME_FreightCostRule, FreightCostRule);
-	}
-
-	/** Get Freight Cost Rule.
-		@return Method for charging Freight
-	  */
-	public String getFreightCostRule () 
-	{
-		return (String)get_Value(COLUMNNAME_FreightCostRule);
-	}
-
-	public org.compiere.model.I_M_FreightCategory getM_FreightCategory() throws RuntimeException
-    {
-		return (org.compiere.model.I_M_FreightCategory)MTable.get(getCtx(), org.compiere.model.I_M_FreightCategory.Table_Name)
-			.getPO(getM_FreightCategory_ID(), get_TrxName());	}
-
-	/** Set Freight Category.
-		@param M_FreightCategory_ID 
-		Category of the Freight
-	  */
-	public void setM_FreightCategory_ID (int M_FreightCategory_ID)
-	{
-		if (M_FreightCategory_ID < 1) 
-			set_Value (COLUMNNAME_M_FreightCategory_ID, null);
-		else 
-			set_Value (COLUMNNAME_M_FreightCategory_ID, Integer.valueOf(M_FreightCategory_ID));
-	}
-
-	/** Get Freight Category.
-		@return Category of the Freight
-	  */
-	public int getM_FreightCategory_ID () 
-	{
-		Integer ii = (Integer)get_Value(COLUMNNAME_M_FreightCategory_ID);
-		if (ii == null)
-			 return 0;
-		return ii.intValue();
-	}
-
-	public org.compiere.model.I_M_Shipper getM_Shipper() throws RuntimeException
-    {
-		return (org.compiere.model.I_M_Shipper)MTable.get(getCtx(), org.compiere.model.I_M_Shipper.Table_Name)
-			.getPO(getM_Shipper_ID(), get_TrxName());	}
-
-	/** Set Shipper.
-		@param M_Shipper_ID 
-		Method or manner of product delivery
-	  */
-	public void setM_Shipper_ID (int M_Shipper_ID)
-	{
-		if (M_Shipper_ID < 1) 
-			set_Value (COLUMNNAME_M_Shipper_ID, null);
-		else 
-			set_Value (COLUMNNAME_M_Shipper_ID, Integer.valueOf(M_Shipper_ID));
-	}
-
-	/** Get Shipper.
-		@return Method or manner of product delivery
-	  */
-	public int getM_Shipper_ID () 
-	{
-		Integer ii = (Integer)get_Value(COLUMNNAME_M_Shipper_ID);
-		if (ii == null)
-			 return 0;
-		return ii.intValue();
-	}
-
-	/** Set Note.
-		@param Note 
-		Optional additional user defined information
-	  */
-	public void setNote (String Note)
-	{
-		set_Value (COLUMNNAME_Note, Note);
-	}
-
-	/** Get Note.
-		@return Optional additional user defined information
-	  */
-	public String getNote () 
-	{
-		return (String)get_Value(COLUMNNAME_Note);
-	}
-
-	/** Set Immutable Universally Unique Identifier.
-		@param UUID 
-		Immutable Universally Unique Identifier
-	  */
-	public void setUUID (String UUID)
-	{
-		set_Value (COLUMNNAME_UUID, UUID);
-	}
-
-	/** Get Immutable Universally Unique Identifier.
-		@return Immutable Universally Unique Identifier
-	  */
-	public String getUUID () 
-	{
-		return (String)get_Value(COLUMNNAME_UUID);
-	}
-
-	/** Set Store Delivery Via.
-		@param W_DeliveryViaRuleAllocation_ID 
-		Delivery Via allowed for Store
-	  */
-	public void setW_DeliveryViaRuleAllocation_ID (int W_DeliveryViaRuleAllocation_ID)
-	{
-		if (W_DeliveryViaRuleAllocation_ID < 1) 
-			set_ValueNoCheck (COLUMNNAME_W_DeliveryViaRuleAllocation_ID, null);
-		else 
-			set_ValueNoCheck (COLUMNNAME_W_DeliveryViaRuleAllocation_ID, Integer.valueOf(W_DeliveryViaRuleAllocation_ID));
-	}
-
-	/** Get Store Delivery Via.
-		@return Delivery Via allowed for Store
-	  */
-	public int getW_DeliveryViaRuleAllocation_ID () 
-	{
-		Integer ii = (Integer)get_Value(COLUMNNAME_W_DeliveryViaRuleAllocation_ID);
-		if (ii == null)
-			 return 0;
-		return ii.intValue();
-	}
-
-	public org.compiere.model.I_W_Store getW_Store() throws RuntimeException
-    {
-		return (org.compiere.model.I_W_Store)MTable.get(getCtx(), org.compiere.model.I_W_Store.Table_Name)
-			.getPO(getW_Store_ID(), get_TrxName());	}
-
-	/** Set Web Store.
-		@param W_Store_ID 
-		A Web Store of the Client
-	  */
-	public void setW_Store_ID (int W_Store_ID)
-	{
-		if (W_Store_ID < 1) 
-			set_ValueNoCheck (COLUMNNAME_W_Store_ID, null);
-		else 
-			set_ValueNoCheck (COLUMNNAME_W_Store_ID, Integer.valueOf(W_Store_ID));
-	}
-
-	/** Get Web Store.
-		@return A Web Store of the Client
-	  */
-	public int getW_Store_ID () 
-	{
-		Integer ii = (Integer)get_Value(COLUMNNAME_W_Store_ID);
-		if (ii == null)
-			 return 0;
-		return ii.intValue();
-	}
-}
