+++ org.spin.store/src/main/java/org/spin/store/model/X_C_PaymentMethod.java
@@ -1,0 +1,213 @@
+/******************************************************************************
+ * Product: ADempiere ERP & CRM Smart Business Solution                       *
+ * Copyright (C) 2006-2017 ADempiere Foundation, All Rights Reserved.         *
+ * This program is free software, you can redistribute it and/or modify it    *
+ * under the terms version 2 of the GNU General Public License as published   *
+ * or (at your option) any later version.										*
+ * by the Free Software Foundation. This program is distributed in the hope   *
+ * that it will be useful, but WITHOUT ANY WARRANTY, without even the implied *
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.           *
+ * See the GNU General Public License for more details.                       *
+ * You should have received a copy of the GNU General Public License along    *
+ * with this program, if not, write to the Free Software Foundation, Inc.,    *
+ * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA.                     *
+ * For the text or an alternative of this public license, you may reach us    *
+ * or via info@adempiere.net or http://www.adempiere.net/license.html         *
+ *****************************************************************************/
+/** Generated Model - DO NOT CHANGE */
+package org.spin.store.model;
+
+import java.sql.ResultSet;
+import java.util.Properties;
+import org.compiere.model.*;
+import org.compiere.util.KeyNamePair;
+
+/** Generated Model for C_PaymentMethod
+ *  @author Adempiere (generated) 
+ *  @version Release 3.9.3 - $Id$ */
+public class X_C_PaymentMethod extends PO implements I_C_PaymentMethod, I_Persistent 
+{
+
+	/**
+	 *
+	 */
+	private static final long serialVersionUID = 20201102L;
+
+    /** Standard Constructor */
+    public X_C_PaymentMethod (Properties ctx, int C_PaymentMethod_ID, String trxName)
+    {
+      super (ctx, C_PaymentMethod_ID, trxName);
+      /** if (C_PaymentMethod_ID == 0)
+        {
+			setC_PaymentMethod_ID (0);
+			setName (null);
+			setTenderType (null);
+			setValue (null);
+        } */
+    }
+
+    /** Load Constructor */
+    public X_C_PaymentMethod (Properties ctx, ResultSet rs, String trxName)
+    {
+      super (ctx, rs, trxName);
+    }
+
+    /** AccessLevel
+      * @return 3 - Client - Org 
+      */
+    protected int get_AccessLevel()
+    {
+      return accessLevel.intValue();
+    }
+
+    /** Load Meta Data */
+    protected POInfo initPO (Properties ctx)
+    {
+      POInfo poi = POInfo.getPOInfo (ctx, Table_ID, get_TrxName());
+      return poi;
+    }
+
+    public String toString()
+    {
+      StringBuffer sb = new StringBuffer ("X_C_PaymentMethod[")
+        .append(get_ID()).append("]");
+      return sb.toString();
+    }
+
+	/** Set Store Payment Method.
+		@param C_PaymentMethod_ID 
+		Payment Methods allowed for Store
+	  */
+	public void setC_PaymentMethod_ID (int C_PaymentMethod_ID)
+	{
+		if (C_PaymentMethod_ID < 1) 
+			set_ValueNoCheck (COLUMNNAME_C_PaymentMethod_ID, null);
+		else 
+			set_ValueNoCheck (COLUMNNAME_C_PaymentMethod_ID, Integer.valueOf(C_PaymentMethod_ID));
+	}
+
+	/** Get Store Payment Method.
+		@return Payment Methods allowed for Store
+	  */
+	public int getC_PaymentMethod_ID () 
+	{
+		Integer ii = (Integer)get_Value(COLUMNNAME_C_PaymentMethod_ID);
+		if (ii == null)
+			 return 0;
+		return ii.intValue();
+	}
+
+	/** Set Description.
+		@param Description 
+		Optional short description of the record
+	  */
+	public void setDescription (String Description)
+	{
+		set_Value (COLUMNNAME_Description, Description);
+	}
+
+	/** Get Description.
+		@return Optional short description of the record
+	  */
+	public String getDescription () 
+	{
+		return (String)get_Value(COLUMNNAME_Description);
+	}
+
+	/** Set Name.
+		@param Name 
+		Alphanumeric identifier of the entity
+	  */
+	public void setName (String Name)
+	{
+		set_Value (COLUMNNAME_Name, Name);
+	}
+
+	/** Get Name.
+		@return Alphanumeric identifier of the entity
+	  */
+	public String getName () 
+	{
+		return (String)get_Value(COLUMNNAME_Name);
+	}
+
+	/** TenderType AD_Reference_ID=214 */
+	public static final int TENDERTYPE_AD_Reference_ID=214;
+	/** Credit Card = C */
+	public static final String TENDERTYPE_CreditCard = "C";
+	/** Check = K */
+	public static final String TENDERTYPE_Check = "K";
+	/** Direct Deposit = A */
+	public static final String TENDERTYPE_DirectDeposit = "A";
+	/** Direct Debit = D */
+	public static final String TENDERTYPE_DirectDebit = "D";
+	/** Account = T */
+	public static final String TENDERTYPE_Account = "T";
+	/** Cash = X */
+	public static final String TENDERTYPE_Cash = "X";
+	/** Credit Memo = M */
+	public static final String TENDERTYPE_CreditMemo = "M";
+	/** Zelle = Z */
+	public static final String TENDERTYPE_Zelle = "Z";
+	/** Mobile Payment Interbank = P */
+	public static final String TENDERTYPE_MobilePaymentInterbank = "P";
+	/** Set Tender type.
+		@param TenderType 
+		Method of Payment
+	  */
+	public void setTenderType (String TenderType)
+	{
+
+		set_Value (COLUMNNAME_TenderType, TenderType);
+	}
+
+	/** Get Tender type.
+		@return Method of Payment
+	  */
+	public String getTenderType () 
+	{
+		return (String)get_Value(COLUMNNAME_TenderType);
+	}
+
+	/** Set Immutable Universally Unique Identifier.
+		@param UUID 
+		Immutable Universally Unique Identifier
+	  */
+	public void setUUID (String UUID)
+	{
+		set_Value (COLUMNNAME_UUID, UUID);
+	}
+
+	/** Get Immutable Universally Unique Identifier.
+		@return Immutable Universally Unique Identifier
+	  */
+	public String getUUID () 
+	{
+		return (String)get_Value(COLUMNNAME_UUID);
+	}
+
+	/** Set Search Key.
+		@param Value 
+		Search key for the record in the format required - must be unique
+	  */
+	public void setValue (String Value)
+	{
+		set_Value (COLUMNNAME_Value, Value);
+	}
+
+	/** Get Search Key.
+		@return Search key for the record in the format required - must be unique
+	  */
+	public String getValue () 
+	{
+		return (String)get_Value(COLUMNNAME_Value);
+	}
+
+    /** Get Record ID/ColumnName
+        @return ID/ColumnName pair
+      */
+    public KeyNamePair getKeyNamePair() 
+    {
+        return new KeyNamePair(get_ID(), getValue());
+    }
+}
